Add files via upload
Code for setting up an excitonic Hamiltonian of a molecular aggregate. The protocol for this calculation is described in the paper "Excitonic Hamiltonians for Calculating Optical Absorption Spectra and Optoelectronic Properties of Molecular Aggregates and Solids" by A.A. Kocherzhenko, S.V. Shedge, P.F. Germaux, M. Heidarian, C.M. Isborn.